Start migrating non-ts files to ts
