--- client/src/App.test.js
@@ -1,8 +1,0 @@
-it('dummy replacement test', () => {  
-});
-
-//it('renders without crashing', () => {
-//  const div = document.createElement('div');
-//  ReactDOM.render(<App />, div);
-//  ReactDOM.unmountComponentAtNode(div);
-//});
