+++ client/src/index.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux'
-import { BrowserRouter } from 'react-router-dom'
-import store from './store'
+import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
+import store from './store';
 import './index.css';
 import App from './components/App';
 import { unregister as unregisterServiceWorker } from './registerServiceWorker';
@@ -13,6 +12,7 @@ ReactDOM.render(
         <BrowserRouter>
             <App />
         </BrowserRouter>
-    </Provider>, 
-document.getElementById('root'));
+    </Provider>,
+    document.getElementById('root')
+);
 unregisterServiceWorker();
